Create fightForladdus.cpp
Problem (Stack)
Given an array, For each element find the value of the nearest element to the right which is having a frequency greater than as that of the current element. If there does not exist an answer for a position, then print '-1'

Please help Raju and his team to solve this problem to get the Laddus.
Input
The first line contains T denoting the no of test cases.
The first line of each test case contains N denoting the no of elements in the array.
The second line of each test case contains N integers (A1..An) denoting the given array.