Vending Machine with JSON
This is a inventory management project for products and their attributes. I have created 2 separate JSON files for the project to be clean for the viewer to read. The two JSON files are Record.json and Sales.json. I also added a seperate program using function to update the added quantity in Items list. For making things a little bit interesting, I also added a program that helps us to calculate one day profit of the respective organization.